slight SEA changes

+++ sea.c
@@ -3,6 +3,7 @@
 #include <archive_entry.h>
 #include <stdlib.h>
 #include <sys/stat.h>
+#include <unistd.h>
 
 extern const char _binary_summit_tar_gz_start[];
 extern const char _binary_summit_tar_gz_end[];
@@ -38,12 +39,17 @@ void extract_mem(const void *data, size_t data_size, const char *output_dir) {
 }
 
 int main(int argc, char* argv[]) {
-    printf("summit SEA (%s %s)\narchive size = %lu bytes\n", __DATE__, __TIME__, _binary_summit_tar_gz_end - _binary_summit_tar_gz_start);
+    if (geteuid() != 0) {
+        fprintf(stderr, "summit requires root permissions to work correctly.\n");
+        return 1;
+    }
+
+    printf("summit SEA (%s %s)\narchive size = %.2f MB\n", __DATE__, __TIME__, (double)(_binary_summit_tar_gz_end - _binary_summit_tar_gz_start)/1000000);
     struct stat st;
     if (stat("/tmp/summit", &st) == 0 && S_ISDIR(st.st_mode)) rmdir("/tmp/summit");
     extract_mem(_binary_summit_tar_gz_start, _binary_summit_tar_gz_end - _binary_summit_tar_gz_start, "/tmp/summit");
 
     if (argc == 1) execv("/tmp/summit/summit-server", (char *[]){"/tmp/summit/summit-server", NULL});
-    else  execv("/tmp/summit/summit-server", (char *[]){"/tmp/summit/summit-server", argv[1], NULL});
+    else execv("/tmp/summit/summit-server", (char *[]){"/tmp/summit/summit-server", argv[1], NULL});
     return 0;
 }
